Moving of services list command to its own package
Having services commands being within the api module makes little sense
if we want to add more services related commands in the future such as
enable/disable.

This would be rather useful
